First commit for many changes. In detail: I started with task 0 completing it modifying the base model files they gave me in the project info.
I organize my repository for the implementation tasks (1 to 4). My intention is to complete correctly the implementation for Login and then if i can for list places.
My main objective although is to complete for Login.

Here i push and save all changes i have been doing these days, the (.html) and (.css) files are not completed for the rest of the webs. Only for Login, the rest may need some changes i will look later. But i make a base of them for task 0.

Now i am with the implementation of Login, recently i coded the Login.js file and the next to do is add de mock-api (API) to the workspace and try to implement it good into my js file.